third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEMorphologyElement-svgdom-operator-prop-expected.txt

SVG 1.1 dynamic update tests
Erode radius = 4

Tests dynamic updates of the 'operator' property of the SVGFEMorphologyElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS morphologyElement.operator.baseVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_DILATE
PASS morphologyElement.operator.baseVal is SVGFEMorphologyElement.SVG_MORPHOLOGY_OPERATOR_ERODE
PASS successfullyParsed is true

TEST COMPLETE
